// src/interfaces/companyInstaller.ts
export interface CompanyInstallerInterface {
  companyName: string;
  phoneNumber: string;
  IMSSNumber: string;
  employeesNumber: number | string;
  website: string;
  ownOffice: boolean | string;
  ownVehicle: boolean | string;
  state: string;
  city: string;
  address: string;
  specializedTools: string;
  certifications: string;
  securityCourses: string;
  yearsExperience: number | string;
}